Move ComposeActivity to it's own package
